refactor : make useSettingDetailPage
