+++ pages/programming/[id].page.tsx
@@ -9,18 +9,14 @@ import { ParsedUrlQuery } from 'querystring';
 import ArticleCollection from 'pages/api/models/articleCollectionModel';
 import DBUtils from 'src/utils/dbUtils';
 
-import { useSetRecoilState, useResetRecoilState } from 'recoil';
-import { detailPageState } from 'src/recoil/article';
-import { ViewArticleElement } from 'src/types/article';
 import { MDXRemote } from 'next-mdx-remote';
+import { useSettingDetailPage } from 'src/hooks';
 
 import MDXTag from 'src/components/article/article-detail/MDXTag';
 import ArticleDetail from 'src/components/article/article-detail/ArticleDetail';
 import ArticleDetailTitle from 'src/components/article/article-detail/ArticleDetailTitle/ArticleDetailTitle';
 import Head from 'src/components/Head';
 import Image from 'next/image';
-
-import { useRouter } from 'next/router';
 
 interface IPrams extends ParsedUrlQuery {
   id: string;
@@ -29,33 +25,7 @@ interface IPrams extends ParsedUrlQuery {
 const Article = ({
   article,
 }: InferGetStaticPropsType<typeof getStaticProps>) => {
-  const setDetailStates = useSetRecoilState(detailPageState);
-  const resetDetailStates = useResetRecoilState(detailPageState);
-  const router = useRouter();
-
-  useEffect(() => {
-    const detailStates: ViewArticleElement = {
-      _id: article._id,
-      title: article.title,
-      content: article.content,
-      imgUrl: article.imgUrl,
-      description: article.description,
-      tags: article.tags,
-      createdAt: article.createdAt,
-      blurDataURL: article.blurDataURL,
-      category: article.category,
-    };
-    setDetailStates(detailStates);
-  }, []);
-
-  useEffect(() => {
-    return () => {
-      router.beforePopState(() => {
-        resetDetailStates();
-        return true;
-      });
-    };
-  }, []);
+  useSettingDetailPage(article);
 
   return (
     <>
