+++ src/hooks/detail-page/useResetDetailPageState.ts
@@ -1,0 +1,17 @@
+import { useEffect } from 'react';
+import { useResetRecoilState } from 'recoil';
+import { detailPageState } from 'src/recoil/article';
+import { useRouter } from 'next/router';
+
+export const useResetDetailPageState = () => {
+  const resetDetailStates = useResetRecoilState(detailPageState);
+  const router = useRouter();
+  useEffect(() => {
+    return () => {
+      router.beforePopState(() => {
+        resetDetailStates();
+        return true;
+      });
+    };
+  }, []);
+};
