+++ src/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useViewportWidth';
 export * from './useArticleElement';
 export * from './article/useResolveSaveArticleHandler';
 export * from './article/useHandleDelete';
+export * from './detail-page/useSettingDetailPage';
